@@ -1,3 +1,3 @@
 module github.com/AlekSi/golang-tip
 
-go 1.20
+go 1.21
